+++ services/app-api/forms/qm.ts
@@ -24,6 +24,7 @@ export const qmReportTemplate: ReportTemplate = {
       title: "General Information",
       type: PageType.Standard,
       sidebar: true,
+      subnavbar: true,
       elements: [
         {
           type: ElementType.Header,
@@ -64,6 +65,7 @@ export const qmReportTemplate: ReportTemplate = {
       title: "Required Measure Results",
       type: PageType.Standard,
       sidebar: true,
+      subnavbar: true,
       elements: [
         {
           type: ElementType.Header,
@@ -84,6 +86,7 @@ export const qmReportTemplate: ReportTemplate = {
       id: "req-measure-result-modal",
       title: "Select measure",
       sidebar: true,
+      subnavbar: true,
       type: PageType.Modal,
       elements: [
         {
@@ -109,6 +112,7 @@ export const qmReportTemplate: ReportTemplate = {
       title: "Stratified Measure Results",
       type: PageType.Standard,
       sidebar: true,
+      subnavbar: true,
       elements: [
         {
           type: ElementType.Header,
@@ -130,6 +134,7 @@ export const qmReportTemplate: ReportTemplate = {
       title: "Review & Submit",
       type: PageType.Standard,
       sidebar: true,
+      subnavbar: true,
       hideNavButtons: true,
       elements: [
         {
