+++ services/app-api/types/reports.ts
@@ -120,6 +120,7 @@ export type FormPageTemplate = {
   type: PageType;
   elements: PageElement[];
   sidebar?: boolean;
+  subnavbar?: boolean;
   hideNavButtons?: boolean;
   childPageIds?: PageId[];
 };
