+++ services/ui-src/src/components/report/ReportPageWrapper.test.tsx
@@ -25,6 +25,7 @@ const testReport: Report = {
       title: "General Information",
       type: PageType.Standard,
       sidebar: true,
+      subnavbar: true,
       elements: [
         {
           type: ElementType.Header,
@@ -43,6 +44,7 @@ const testReport: Report = {
       title: "Required Measure Results",
       type: PageType.Standard,
       sidebar: true,
+      subnavbar: true,
       elements: [
         {
           type: ElementType.Header,
